backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asgiref==3.9.1",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "Django==5.2.6",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "google-auth==2.40.3",
    "google-genai==1.33.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "pillow==11.3.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "PyJWT==2.10.1",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tenacity==9.1.2",
    "typing-inspection==0.4.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "websockets==15.0.1",
    "channels==4.0.0",
    "channels-redis==4.1.0",
    "daphne==4.0.0",
    "redis==5.0.1",
    "django-redis==5.4.0",
    "hiredis==2.3.2",
    "pyfcm==1.5.4",
]

[tool.uv]
